+++ .storybook-js/main.js
@@ -1,4 +1,5 @@
-module.exports = {
+export default {
   addons: ['@storybook/addon-actions'],
   stories: ['../packages/autocomplete-js/Autocomplete.stories.js'],
+  framework: '@storybook/html-vite',
 }