+++ .storybook-vue/preview.js
@@ -1,21 +1,14 @@
-import Vue from 'vue'
-import { addDecorator, addParameters } from '@storybook/vue';
-import { withStyles } from 'storybook-addon-styles/vue'
 import '../packages/style.css'
-import './style.css'
-import Autocomplete from '../packages/autocomplete-vue/Autocomplete.vue'
 
-Vue.component('Autocomplete', Autocomplete)
-
-addDecorator(withStyles)
-addParameters({
-  options: {
-    showPanel: false
+const preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
   },
-  styles: {
-    margin: '0 auto',
-    padding: '40px 24px 0',
-    maxWidth: '400px',
-    fontFamily: '-apple-system, BlinkMacSystemFont, "Segoe UI", "Helvetica Neue", sans-serif',
-  }
-})
+}
+
+export default preview