+++ babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  presets: ['@babel/env'],
-  plugins: ['@babel/proposal-class-properties'],
-}