+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v3.0.0 (March 28, 2024)
+
+- Updated vue component for Vue.js 3.x
+- Updated storybook to v8.0. Updated stories to new format.
+- Updated dev dependencies
+- Updated node.js version to 18 (newer versions untested)
+
 ## v2.4.1 (January 13, 2023)
 
 ### autocomplete-core